pyproject.toml
[project]
name = "juliet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "openai>=1.109.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]
